2020.10.27-oss-opentelemetry/requirements.txt
opentelemetry-api==0.13b0
opentelemetry-exporter-jaeger==0.13b0
opentelemetry-instrumentation==0.13b0
opentelemetry-instrumentation-dbapi==0.13b0
opentelemetry-instrumentation-flask==0.13b0
opentelemetry-instrumentation-mysql==0.13b0
opentelemetry-instrumentation-requests==0.13b0
opentelemetry-instrumentation-wsgi==0.13b0
opentelemetry-sdk==0.13b0
